% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wear_time.R
\name{calculate_wear_time}
\alias{calculate_wear_time}
\title{calculate_wear_time}
\usage{
calculate_wear_time(directory, valid_wear_time = c(8, 10))
}
\arguments{
\item{directory}{Path name to the location of the .agd or .csv files}

\item{valid_wear_time}{Number of hours to be considered a valid day, Default: c(8, 10)}
}
\value{
A data frame containing the number of valid days,
}
\description{
Calculates wear time for several .agd or .csv files
}
\details{
Calculates wear time for several .agd or .csv files
}
\seealso{
\code{\link[dplyr]{rename}}
\code{\link[PhysicalActivity]{wearingMarking}}
\code{\link[plyr]{rbind.fill}}
}
