% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ilr_transform.R
\name{get_ilr}
\alias{get_ilr}
\title{Isometric Log Ratio Transformation with a Sequential Binary Partition}
\usage{
get_ilr(X, sbp)
}
\arguments{
\item{X}{Matrix of data frame that includes only variables in the composition}

\item{sbp}{A matrix containing the sequential binary partition}
}
\value{
A data frame with (n-1) columns containing the isometric log ratio
transformed values.
}
\description{
This function does a isometric log ratio transformation using
methods similar to the \code{compositions} R package.
}
\details{
This function does an isometric log ratio transformation as described
by \href{https://www.sciencedirect.com/science/article/pii/S009830040700101X}{van den Boogaart et al., 2008}
and comparable to the \code{ilr} function from the \code{compositions} package.
}
