% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ilr_transform.R
\name{ilr_pivotCoord}
\alias{ilr_pivotCoord}
\title{Pivot coordinates for Isometric Log Ratio Transformation}
\usage{
ilr_pivotCoord(X)
}
\arguments{
\item{X}{Matrix or data frame that includes only variables in the composition.}
}
\value{
A data frame with (n-1) columns containing the isometric log ratio
transformed values.
}
\description{
This function does an isometric log ratio transformation using
methods similar to the \code{robCompositions} R package.
}
\details{
This function does an isometric log ratio transformation as
described by \href{https://pubmed.ncbi.nlm.nih.gov/26461112/}{Chastin et al., 2005}
and comparable to the \code{pivotCoord} function from the \code{robCompositions} package.
}
