% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.R
\name{mars.main}
\alias{mars.main}
\title{Movement Analysis for Remote Sensors (MARS)}
\usage{
mars.main(
  study.name = "study",
  datadir,
  results,
  adult.cp = "",
  child.cp = "",
  spurious = 20000,
  axis = 1,
  overwrite = FALSE,
  person.time = FALSE,
  person.date = TRUE,
  person.month = FALSE,
  valid_time = NULL,
  valid_days = NULL,
  valid_weekend = NULL,
  return.timestamped.dataframe = FALSE,
  individual.file.save = FALSE,
  do.parallel = TRUE
)
}
\arguments{
\item{study.name}{Abbreviation of the Study, Default: study}

\item{datadir}{Directory where the AGD files are stored.}

\item{results}{Directory where the results should be stored.}

\item{adult.cp}{Adult cut-point to be used (see cutpoint.list.R for list of available cut-points), Default: ''}

\item{child.cp}{Child cut-point to be used (see cutpoint.list.R for list of available cut-points), Default: ''}

\item{spurious}{A maximum value for vertical or vector magnitude counts to not exceed, Default: 20000}

\item{axis}{Number of axes to be used (1: Vertical Axis; 3: Vector Magnitude), Default: 1}

\item{overwrite}{Overwrite the individual files if individual.file.save is set to TRUE, Default: FALSE}

\item{person.time}{Export the accelerometer summary by person time categories, Default: FALSE}

\item{person.date}{Export the accelerometer summary by person date, Default: TRUE}

\item{person.month}{Export only valid days to a summary by person time point of the trial, Default: FALSE}

\item{valid_time}{Number of hours or minutes for a valid day (converts values between 1-24 to minutes by multiplying by 60), Default: NULL}

\item{valid_days}{Number of days for inclusion, Default: NULL}

\item{valid_weekend}{Number of weekend days for inclusion, Default: NULL}

\item{return.timestamped.dataframe}{Return a timestamped data set into R for all combined time point and id data, Default: FALSE}

\item{individual.file.save}{Save all of the individual timestamped files to a CSV, Default: FALSE}

\item{do.parallel}{Process dates in parallel to speed up the processing of the accelerometer data}
}
\value{
Individual and/or summary files to a Results folder located in the designated directory.
}
\description{
Main function to execute MARS accelerometer processing program.
}
\details{
Main function to execute MARS accelerometer processing program.
}
\seealso{
\code{\link[readr]{write_delim}}
\code{\link[plyr]{rbind.fill}}
}
