% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoveKC.R
\docType{_PACKAGE}
\name{MoveKC}
\alias{MoveKC}
\title{MoveKC: A package for physical activity and weight management research.}
\description{
The MoveKC package provides several important functions:

\code{\link{ilr_pivotCoord}}

\code{\link{get_ilr}}

\code{\link{mars.main}}

\code{\link{read_agd}}

\code{\link{agd_to_csv}}

\code{\link{gt3x2csv}}

\code{\link{AGread.csv}}

\code{\link{birth.date}}

\code{\link{AG.temporal}}

\code{\link{cutpoints}}

\code{\link{detect.bouts}}
}
